Improve osm extraction Fix bug with the SRID when ST_Union is used Refactoring the unit tests for the processing chain
