+++ prepare_data/src/test/groovy/org/orbisgis/osm/FormattingForAbstractModelTests.groovy
@@ -12,92 +12,65 @@ import static org.junit.jupiter.api.Assertions.*
 class FormattingForAbstractModelTests {
 
     @Test
-   void transformBuildingsTest() {
+   void formatingGISLayers() {
         def h2GIS = H2GIS.open('./target/osmdb;AUTO_SERVER=TRUE')
         IProcess extractData = PrepareData.OSMGISLayers.createGISLayers()
         extractData.execute([
                 datasource : h2GIS,
                 osmFilePath: new File(this.class.getResource("redon.osm").toURI()).getAbsolutePath(),
                 epsg :2154])
-        assertEquals 1038, h2GIS.getTable(extractData.results.buildingTableName).rowCount
 
+        assertEquals 1038, h2GIS.getTable(extractData.results.buildingTableName).rowCount
+        assertEquals 360, h2GIS.getTable(extractData.results.roadTableName).rowCount
+        assertEquals 44, h2GIS.getTable(extractData.results.railTableName).rowCount
+        assertEquals 128, h2GIS.getTable(extractData.results.vegetationTableName).rowCount
+        assertEquals 8, h2GIS.getTable(extractData.results.hydroTableName).rowCount
+
+
+        //Buildings
         IProcess format = PrepareData.FormattingForAbstractModel.formatBuildingLayer()
         format.execute([
                 datasource : h2GIS,
                 inputTableName: extractData.results.buildingTableName])
         assertEquals 1038, h2GIS.getTable(format.results.outputTableName).rowCount
         h2GIS.getTable(format.results.outputTableName).save("./target/osm_building_formated.shp")
-    }
 
-    @Test
-    void transformRoadsTest() {
-        def h2GIS = H2GIS.open('./target/osmdb;AUTO_SERVER=TRUE')
-        IProcess extractData = PrepareData.OSMGISLayers.createGISLayers()
-        extractData.execute([
-                datasource : h2GIS,
-                osmFilePath: new File(this.class.getResource("redon.osm").toURI()).getAbsolutePath(),
-                epsg :2154])
-        assertEquals 360, h2GIS.getTable(extractData.results.roadTableName).rowCount
-
-        IProcess format = PrepareData.FormattingForAbstractModel.formatRoadLayer()
+        //Roads
+        format = PrepareData.FormattingForAbstractModel.formatRoadLayer()
         format.execute([
                 datasource : h2GIS,
                 inputTableName: extractData.results.roadTableName])
 
         assertEquals 360, h2GIS.getTable(format.results.outputTableName).rowCount
         h2GIS.getTable(format.results.outputTableName).save("./target/osm_road_formated.shp")
-    }
 
-    @Test
-    void transformRailsTest() {
-        def h2GIS = H2GIS.open('./target/osmdb;AUTO_SERVER=TRUE')
-        IProcess extractData = PrepareData.OSMGISLayers.createGISLayers()
-        extractData.execute([
-                datasource : h2GIS,
-                osmFilePath: new File(this.class.getResource("redon.osm").toURI()).getAbsolutePath(),
-                epsg :2154])
-        assertEquals 44, h2GIS.getTable(extractData.results.railTableName).rowCount
+        //Rails
 
-        IProcess format = PrepareData.FormattingForAbstractModel.formatRailsLayer()
+        format = PrepareData.FormattingForAbstractModel.formatRailsLayer()
         format.execute([
                 datasource : h2GIS,
                 inputTableName: extractData.results.railTableName])
 
         assertEquals 44, h2GIS.getTable(format.results.outputTableName).rowCount
         h2GIS.getTable(format.results.outputTableName).save("./target/osm_rails_formated.shp")
-    }
 
+        //Vegetation
 
-    @Test
-    void transformVegetationTest() {
-        def h2GIS = H2GIS.open('./target/osmdb;AUTO_SERVER=TRUE')
-        IProcess extractData = PrepareData.OSMGISLayers.createGISLayers()
-        extractData.execute([
-                datasource : h2GIS,
-                osmFilePath: new File(this.class.getResource("redon.osm").toURI()).getAbsolutePath(),
-                epsg :2154])
-        assertEquals 128, h2GIS.getTable(extractData.results.vegetationTableName).rowCount
-
-        IProcess format = PrepareData.FormattingForAbstractModel.formatVegetationLayer()
+        format = PrepareData.FormattingForAbstractModel.formatVegetationLayer()
         format.execute([
                 datasource : h2GIS,
                 inputTableName: extractData.results.vegetationTableName])
 
         assertEquals 128, h2GIS.getTable(format.results.outputTableName).rowCount
+        assertTrue h2GIS.firstRow("select count(*) as count from ${format.results.outputTableName} where type is null").count==0
+
+        assertTrue h2GIS.firstRow("select count(*) as count from ${format.results.outputTableName} where HEIGHT_CLASS is null").count==0
+
         h2GIS.getTable(format.results.outputTableName).save("./target/osm_vegetation_formated.shp")
-    }
 
-    @Test
-    void transformHydroTest() {
-        def h2GIS = H2GIS.open('./target/osmdb;AUTO_SERVER=TRUE')
-        IProcess extractData = PrepareData.OSMGISLayers.createGISLayers()
-        extractData.execute([
-                datasource : h2GIS,
-                osmFilePath: new File(this.class.getResource("redon.osm").toURI()).getAbsolutePath(),
-                epsg :2154])
-        assertEquals 8, h2GIS.getTable(extractData.results.hydroTableName).rowCount
+        //Hydrography
 
-        IProcess format = PrepareData.FormattingForAbstractModel.formatVegetationLayer()
+        format = PrepareData.FormattingForAbstractModel.formatHydroLayer()
         format.execute([
                 datasource : h2GIS,
                 inputTableName: extractData.results.hydroTableName])
@@ -105,6 +78,76 @@ class FormattingForAbstractModelTests {
         assertEquals 8, h2GIS.getTable(format.results.outputTableName).rowCount
         h2GIS.getTable(format.results.outputTableName).save("./target/osm_hydro_formated.shp")
     }
+
+    //@Test //enable it to test data extraction from the overpass api
+    void extractCreateFormatGISLayers() {
+        def h2GIS = H2GIS.open('./target/osmdb;AUTO_SERVER=TRUE')
+
+        //def placeName ="Shanghai, Chine"
+        def placeName ="Rennes"
+
+        IProcess extractData = PrepareData.OSMGISLayers.extractAndCreateGISLayers()
+        extractData.execute([
+                datasource : h2GIS,
+                placeName:placeName ])
+
+        String formatedPlaceName = placeName.trim().split("\\s*(,|\\s)\\s*").join("_");
+
+
+        if(extractData.results.zoneTableName!=null) {
+            //Zone
+            h2GIS.getTable(extractData.results.zoneTableName).save("./target/osm_zone_${formatedPlaceName}.geojson")
+
+            //Zone envelope
+            h2GIS.getTable(extractData.results.zoneEnvelopeTableName).save("./target/osm_zone_envelope_${formatedPlaceName}.geojson")
+
+
+            //Buildings
+            IProcess format = PrepareData.FormattingForAbstractModel.formatBuildingLayer()
+            format.execute([
+                    datasource : h2GIS,
+                    inputTableName: extractData.results.buildingTableName])
+            h2GIS.getTable(format.results.outputTableName).save("./target/osm_building_${formatedPlaceName}.geojson")
+
+            //Roads
+            format = PrepareData.FormattingForAbstractModel.formatRoadLayer()
+            format.execute([
+                    datasource : h2GIS,
+                    inputTableName: extractData.results.roadTableName])
+            h2GIS.getTable(format.results.outputTableName).save("./target/osm_road_${formatedPlaceName}.geojson")
+
+
+            //Rails
+            format = PrepareData.FormattingForAbstractModel.formatRailsLayer()
+            format.execute([
+                    datasource : h2GIS,
+                    inputTableName: extractData.results.railTableName])
+            h2GIS.getTable(format.results.outputTableName).save("./target/osm_rails_${formatedPlaceName}.geojson")
+
+
+            //Vegetation
+            format = PrepareData.FormattingForAbstractModel.formatVegetationLayer()
+            format.execute([
+                    datasource : h2GIS,
+                    inputTableName: extractData.results.vegetationTableName])
+            h2GIS.getTable(format.results.outputTableName).save("./target/osm_vegetation_${formatedPlaceName}.geojson")
+
+
+            //Hydrography
+            format = PrepareData.FormattingForAbstractModel.formatHydroLayer()
+            format.execute([
+                    datasource : h2GIS,
+                    inputTableName: extractData.results.hydroTableName])
+            h2GIS.getTable(format.results.outputTableName).save("./target/osm_hydro_${formatedPlaceName}.geojson")
+
+        }else {
+            assertTrue(false)
+        }
+    }
+
+
+
+
 }
