+++ prepare_data/src/test/groovy/org/orbisgis/osm/OSMGISLayersTests.groovy
@@ -14,7 +14,7 @@ class OSMGISLayersTests {
 
     private static final Logger logger = LoggerFactory.getLogger(OSMGISLayersTests.class)
 
-    @Test //disable due to potential API blocking
+    //@Test enable it to test data extraction from the overpass api
     void extractAndCreateGISLayers() {
         def h2GIS = H2GIS.open('./target/osmdb;AUTO_SERVER=TRUE')
         IProcess process = PrepareData.OSMGISLayers.extractAndCreateGISLayers()
